Rework benchmark driver scripts.
 * Introduce getopts option parsing to all scripts.
 * Add ability to specify benchmark specific arguments.
 * Benchmark scripts persist pid information and support
   waiting for their completion via -w flag.
 * Benchmark scripts now run under tmux session on the remote
   server so that transient issues do not interrupt benchmarks;
   Allows ssh-ing and attaching to the session.
 * Start micro benchmarks in parallel on different nodes.

+++ scripts/gen/cpu.sh
@@ -1,11 +1,45 @@
 #!/bin/bash
 
+set -ex
+pidfile="$HOME/cpu-bench.pid"
+f_force=''
+f_wait=''
+f_iters=10
+
+while getopts 'fwn:' flag; do
+  case "${flag}" in
+    f) f_flag='true' ;;
+    n) f_iters="${OPTARG}" ;;
+    w) f_wait='true' ;;
+    *) echo "Usage: $0 [-f] [-w] [-n num_iterations]"
+       exit 1 ;;
+  esac
+done
+
 logdir="$HOME/coremark-results"
+
+if [ -n "$f_wait" ];
+then
+  exec sh -c "
+    ( test -f '$logdir/success' ||
+      (tail --pid \$(cat $pidfile) -f /dev/null && test -f '$logdir/success')
+    ) || (echo 'CPU benchmark did not complete successfully.  Check logs'; exit 1)"
+fi
+
+if [ -f "$pidfile" ] && [ -z "$f_force" ] ;
+then
+  pid=$(cat $pidfile)
+  echo "CPU benchmark already running (pid $pid)"
+  exit
+fi
+
+trap "rm -f $pidfile" EXIT SIGINT
+echo $$ > "$pidfile"
+
 rm -rf "$logdir"
 mkdir "$logdir"
 
 exec &> >(tee "$logdir/script.log")
-set -ex
 
 if [ ! -d coremark ]
 then
@@ -16,9 +50,20 @@ cd coremark
 
 # Build default coremark (single proc)
 make REBUILD=1 link
-./coremark.exe > "${logdir}/single.log"
+
+for ((i=0; i < f_iters; i++))
+do
+  echo "Single core iteration: $i"
+  ./coremark.exe > "${logdir}/single-$i.log"
+done
+
 
 # Rebuild to run in multithreaded mode.
 make LFLAGS_END="-lpthread" XCFLAGS="-DMULTITHREAD=$(nproc) -DUSE_PTHREAD" REBUILD=1 link
-./coremark.exe > "${logdir}/multi.log"
+for ((i=0; i < f_iters; i++))
+do
+  echo "Multi core iteration: $i"
+  ./coremark.exe > "${logdir}/multi-$i.log"
+done
 
+touch "$logdir/success"
